refactor: splitt primaryAllocation.ts etter subdomener (#173)
2573-linjersfilen splittet i 7 domenefiler under primaryAllocation/:
- _helpers.ts: delte private hjelpefunksjoner og typer
- costs.ts: kostnadsberegning og pool-forespørsler
- access.ts: profesjon/samfunns-tilgangskontroll
- draftState.ts: draft state-mutasjoner og spend/remove-operasjoner
- views.ts: buildChargenDraftView og view-hjelpere
- review.ts: reviewChargenDraft
- finalize.ts: finalizeChargenDraft

primaryAllocation.ts er nå et kompatibelt re-eksport-barrel.
Alle 230 tester grønne.
